+++ src/common/mpeg1_2.h
@@ -62,7 +62,7 @@ namespace mpeg1_2 {
 
 int extract_fps_idx(const unsigned char *buffer, int buffer_size);
 double get_fps(int idx);
-bool extract_ar(const unsigned char *buffer, int buffer_size, float &ar);
+bool extract_ar(const unsigned char *buffer, int buffer_size, double &ar);
 bool is_fourcc(uint32_t fourcc);
 
 };
