+++ src/input/r_qtmp4.h
@@ -264,7 +264,7 @@ struct qtmp4_demuxer_c {
   bool m_hevc_is_annex_b{};
   std::deque<int64_t> references;
   uint32_t a_channels, a_bitdepth;
-  float a_samplerate;
+  double a_samplerate;
   boost::optional<mtx::aac::audio_config_t> a_aac_audio_config;
   mtx::ac3::frame_c m_ac3_header;
   mtx::dts::header_t m_dts_header;
