+++ src/output/p_flac.cpp
@@ -58,7 +58,7 @@ void
 flac_packetizer_c::set_headers() {
   set_codec_id(MKV_A_FLAC);
   set_codec_private(m_header);
-  set_audio_sampling_freq((float)m_stream_info.sample_rate);
+  set_audio_sampling_freq(m_stream_info.sample_rate);
   set_audio_channels(m_stream_info.channels);
   set_audio_bit_depth(m_stream_info.bits_per_sample);
